Split Server.Endpoints.Tx.Class to 2 modules
Move things that are irrelevant to the HasTxEndpoints class into a separate module
