+++ src/Server/Endpoints/Tx/Class.hs
@@ -1,9 +1,5 @@
-{-# LANGUAGE AllowAmbiguousTypes        #-}
 {-# LANGUAGE DataKinds                  #-}
-{-# LANGUAGE DeriveGeneric              #-}
-{-# LANGUAGE DerivingVia                #-}
 {-# LANGUAGE FlexibleContexts           #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE TypeFamilies               #-}
 {-# LANGUAGE UndecidableInstances       #-}
 {-# LANGUAGE UndecidableSuperClasses    #-}
@@ -12,13 +8,12 @@ module Server.Endpoints.Tx.Class where
 
 import           Control.Monad.Catch              (Exception)
 import           Control.Monad.Reader             (MonadReader)
-import           Data.Aeson                       (ToJSON)
 import           Data.Kind                        (Type)
 import           Data.Text                        (Text)
-import           GHC.Generics                     (Generic)
 import           IO.Wallet                        (HasWallet(..))
-import           Servant                          (NoContent(..), Union, IsMember, WithStatus, HasStatus)
+import           Servant                          (NoContent(..), Union, IsMember, WithStatus)
 import           Server.Class                     (NetworkM, HasServer(..), Env)
+import           Server.Endpoints.Tx.Internal     (NewTxEndpointResult)
 import           Types.Tx                         (TransactionBuilder)
 
 class ( HasServer s
@@ -38,10 +33,3 @@ class ( HasServer s
     checkForTxEndpointsErrors :: InputOf s -> NetworkM s ()
 
     txEndpointsErrorHandler :: TxEndpointsErrorOf s -> NetworkM s (Union (TxApiResultOf s))
-
-type DefaultTxApiResult = '[WithStatus 422 Text, NoContent, NewTxEndpointResult]
-
-newtype NewTxEndpointResult = NewTxEndpointResult Text
-    deriving HasStatus via WithStatus 200 NewTxEndpointResult
-    deriving (Show, Generic)
-    deriving newtype ToJSON
