+++ src/Server/Endpoints/Tx/New.hs
@@ -10,7 +10,8 @@ module Server.Endpoints.Tx.New where
 import           Control.Monad                    (join, liftM3)
 import           Control.Monad.Catch              (handle)
 import           Servant                          (JSON, (:>), ReqBody, respond, StdMethod(POST), UVerb, Union)
-import           Server.Endpoints.Tx.Class        (HasTxEndpoints(..), NewTxEndpointResult(..))
+import           Server.Endpoints.Tx.Class        (HasTxEndpoints(..))
+import           Server.Endpoints.Tx.Internal     (NewTxEndpointResult(..))
 import           Server.Class                     (NetworkM, HasServer(..))
 import           Server.Tx                        (mkBalanceTx)
 import           Utils.Logger                     (HasLogger(..), (.<))
