+++ src/TestingServer/Main.hs
@@ -20,7 +20,8 @@ import Data.List                    (nub)
 import Options.Applicative          (argument, metavar, str)
 import Plutus.V2.Ledger.Api         (BuiltinByteString)
 import PlutusTx.Builtins.Class      (stringToBuiltinByteString)
-import Server.Endpoints.Tx.Class    (HasTxEndpoints(..), DefaultTxApiResult)
+import Server.Endpoints.Tx.Class    (HasTxEndpoints(..))
+import Server.Endpoints.Tx.Internal (DefaultTxApiResult)
 import Server.Class                 (HasServer(..))
 import System.Random                (randomRIO, randomIO)
 import TestingServer.OffChain       (testMintTx)
